refactor: turn useProposal hook into context
