+++ src/utils/proposals.ts
@@ -5,6 +5,10 @@ import {
 } from './index';
 import moment from 'moment';
 
+import DaoStore from 'stores/DaoStore';
+import { ZERO_ADDRESS } from 'utils';
+import { ProposalsExtended } from 'contexts/proposals';
+
 // constant used to the initial order of the proposals (Any Status).
 export const QUEUED_PRIORITY_THRESHOLD = 10;
 
@@ -349,3 +353,77 @@ export const decodeProposalStatus = function (
 //is it added in Cache?
 export const orderByNewestTimeToFinish = (a: any, b: any) =>
   a.finishTime - b.finishTime;
+
+/// filtering and sorting proposals for All States, All Schemas criteria
+export const filterInitialCriteria = (
+  proposals: ProposalsExtended[],
+  daoStore: DaoStore
+) => {
+  // (QuitedEndingPeriod || Queded) && positiveVotes >= 10% (Ordered from time to finish, from lower to higher)
+  let earliestAbove10 = proposals.filter((proposal: Proposal) => {
+    const repAtCreation = daoStore.getRepAt(
+      ZERO_ADDRESS,
+      proposal.creationEvent.l1BlockNumber
+    ).totalSupply;
+
+    return (
+      (proposal.stateInVotingMachine ===
+        VotingMachineProposalState.QuietEndingPeriod ||
+        proposal.stateInVotingMachine === VotingMachineProposalState.Queued) &&
+      proposal.positiveVotes
+        .div(repAtCreation)
+        .times(100)
+        .decimalPlaces(2)
+        .gte(QUEUED_PRIORITY_THRESHOLD)
+    );
+  });
+  earliestAbove10.sort(orderByNewestTimeToFinish);
+
+  // Proposals Boosted. (Ordered from time to finish, from lower to higher)
+  let boosted = proposals.filter(
+    (proposal: Proposal): Boolean =>
+      proposal.stateInVotingMachine === VotingMachineProposalState.Boosted
+  );
+  boosted.sort(orderByNewestTimeToFinish);
+
+  let preBoosted = proposals.filter(
+    (proposal: Proposal): Boolean =>
+      proposal.stateInVotingMachine === VotingMachineProposalState.PreBoosted
+  );
+  preBoosted.sort(orderByNewestTimeToFinish);
+
+  // (QuitedEndingPeriod || Queded) && positiveVotes < 10% (Ordered from time to finish, from lower to higher)
+  let earliestUnder10 = proposals.filter((proposal: Proposal): Boolean => {
+    const repAtCreation = daoStore.getRepAt(
+      ZERO_ADDRESS,
+      proposal.creationEvent.l1BlockNumber
+    ).totalSupply;
+
+    return (
+      (proposal.stateInVotingMachine ===
+        VotingMachineProposalState.QuietEndingPeriod ||
+        proposal.stateInVotingMachine === VotingMachineProposalState.Queued) &&
+      proposal.positiveVotes
+        .div(repAtCreation)
+        .times(100)
+        .decimalPlaces(2)
+        .lt(QUEUED_PRIORITY_THRESHOLD)
+    );
+  });
+  earliestUnder10.sort(orderByNewestTimeToFinish);
+
+  //Proposals in Executed status. (Ordered in time passed since finish, from lower to higher)
+  let executed = proposals.filter(
+    (proposal: Proposal): Boolean =>
+      proposal.stateInVotingMachine === VotingMachineProposalState.Executed
+  );
+  executed.sort(orderByNewestTimeToFinish);
+
+  return [
+    ...earliestAbove10,
+    ...boosted,
+    ...preBoosted,
+    ...earliestUnder10,
+    ...executed,
+  ];
+};
